# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Clínica Serena API
  description: |
    Contrato oficial de la API de Clínica Serena. Es la fuente de verdad
    compartida entre el backend (Spring Boot), el frontend web (Next.js) y
    la app Android del paciente.

    Define los endpoints iniciales, agrupados en Health, Catálogo, Auth,
    Citas, Portal del paciente y Personal (ver tags).

    ## Estado de implementación
    - **Implementado en esta fase**: entidades y migración de paciente, cuenta
      y sesión; `POST /auth/login`, `GET /auth/me` y `POST /auth/logout` con
      BCrypt y Bearer opaco revocable.
    - **Implementado en esta fase**: autenticación Bearer separada para
      personal, creación administrativa de cuentas de recepción/médico,
      agenda persistida y registro de llegada con auditoría.
    - **Implementado para acceso web unificado**: `POST /auth/login-unified`
      resuelve la cuenta por correo y contraseña sin aceptar un rol del cliente.
      Si el mismo correo existe en paciente y personal, rechaza el acceso de
      forma genérica hasta que esa ambigüedad sea resuelta administrativamente.
    - **Implementado en esta fase**: vinculación administrativa de cuentas
      MEDICO con profesionales del catálogo, citas propias del profesional,
      expediente básico, historial de atenciones y registro inmutable de
      atención con diagnóstico y receta (`/medico/**`).
    - **Implementado para el portal del paciente**: perfil propio consultable y
...

...
servers:
...
paths:
...
  /staff/especialidades:
    get:
      tags: [Personal]
      summary: Lista especialidades para ADMIN
      description: Requiere rol ADMIN. Es la vista administrativa del mismo catálogo persistido que consume el endpoint público.
      security:
        - bearerAuth: []
      operationId: listAdminSpecialties
      responses:
        "200":
          description: Especialidades persistidas.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Specialty" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Personal]
      summary: Crea una especialidad
      description: Requiere ADMIN. El servidor recorta, colapsa espacios y normaliza el nombre a minúsculas antes de persistirlo. No recibe categoría porque el modelo actual no la persiste.
      security:
        - bearerAuth: []
      operationId: createAdminSpecialty
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertSpecialtyRequest" }
      responses:
        "201":
          description: Especialidad creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Specialty" }
        "400":
          description: Nombre o descripción vacíos o fuera de longitud.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Ya existe una especialidad con el mismo nombre normalizado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /staff/especialidades/{specialtyId}:
    patch:
      tags: [Personal]
      summary: Edita una especialidad
      description: Requiere ADMIN. Conserva el identificador y las relaciones existentes con médicos y citas; no existe eliminación física.
      security:
        - bearerAuth: []
      operationId: updateAdminSpecialty
      parameters:
        - name: specialtyId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertSpecialtyRequest" }
      responses:
        "200":
          description: Especialidad actualizada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Specialty" }
        "400":
          description: Nombre o descripción inválidos.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: La especialidad no existe.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: El nuevo nombre colisiona con otra especialidad.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /staff/medico/horarios:
    get:
      tags: [Personal]
      summary: Consulta los bloques propios del médico
      description: Requiere MEDICO con vinculación activa. El profesional se deriva de la cuenta autenticada; no se acepta medicoId.
      security:
        - bearerAuth: []
      operationId: listOwnMedicalScheduleBlocks
      responses:
        "200":
          description: Bloques propios ordenados por inicio, incluidos los ya reservados.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ScheduleBlock" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/MedicalForbidden" }
    post:
      tags: [Personal]
      summary: Agrega un bloque propio
      description: |
        Requiere MEDICO con vinculación activa. `startAt` y `endAt` deben
        representar la hora de pared de `America/Guatemala` (UTC-06:00) cuando
        provienen de la UI. Valida inicio/fin, rechaza un `startAt` pasado y
        solapamientos, y serializa altas del mismo profesional con un bloqueo
        pesimista. Agregar un bloque no modifica citas existentes.
      security:
        - bearerAuth: []
      operationId: addOwnMedicalScheduleBlock
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleBlockRequest" }
      responses:
        "201":
          description: Bloque disponible creado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScheduleBlock" }
        "400":
          description: Inicio/fin ausente, formato inválido, inicio no anterior al fin o `startAt` ya pasado (`SCHEDULE_START_IN_PAST`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/MedicalForbidden" }
        "409":
          description: El bloque se solapa con otro del mismo profesional.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
  /staff/medico/horarios/{blockId}:
    patch:
      tags: [Personal]
      summary: Edita un bloque futuro propio sin citas asociadas
      security: [{ bearerAuth: [] }]
      parameters: [{ name: blockId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ScheduleBlockRequest" } } } }
      responses:
        "200": { description: Bloque actualizado. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { description: No es el médico propietario. }
        "404": { description: Bloque inexistente. }
        "409": { description: Fue retirado, tiene citas que impiden editar o se solapa. }
        "422": { description: El bloque actual o el nuevo inicio no es futuro, o la duración es inválida. }
    delete:
      tags: [Personal]
      summary: Retira lógicamente un bloque futuro propio conservando sus citas
      description: El bloqueo pesimista serializa el retiro con reservas. Marca el bloque no disponible sin borrar ni modificar citas existentes.
      security: [{ bearerAuth: [] }]
      parameters: [{ name: blockId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "204": { description: Retirado sin borrar el registro. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { description: No es el médico propietario. }
        "404": { description: Bloque inexistente. }
        "422": { description: Ya no es futuro. }
  /staff/patients/{patientId}/link-requests:
    post:
      tags: [Personal]
      summary: Inicia una vinculación tras verificar identidad presencialmente
      description: No recibe ni busca cuentas por nombre, correo o teléfono. El código se muestra una sola vez, expira en diez minutos y solo se almacena su hash.
      security: [{ bearerAuth: [] }]
      parameters: [{ name: patientId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [identityVerifiedInPerson], properties: { identityVerifiedInPerson: { type: boolean, enum: [true] } } } } }
      responses:
        "201": { description: Desafío creado. }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: Expediente inexistente. }
        "409": { description: Expediente ya vinculado. }
        "422": { description: Falta verificación presencial. }
  /pacientes/me/administrative-link:
    post:
      tags: [Pacientes]
      summary: Confirma una vinculación con la cuenta autenticada
      description: La cuenta objetivo proviene de PatientPrincipal. El código es de un uso, tiene cinco intentos y nunca se registra.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [requestId, code], properties: { requestId: { type: string, format: uuid }, code: { type: string, pattern: "^[0-9]{8}$" } } } } }
      responses:
        "204": { description: Vinculación confirmada. }
        "404": { description: Solicitud inexistente. }
        "409": { description: No vigente, duplicada o expediente con dependencias. }
        "410": { description: Código vencido. }
        "422": { description: Código incorrecto. }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Token opaco emitido por `POST /auth/login`. Solo se almacena su hash.
        Expira en 1800 segundos por defecto y se revoca con `POST /auth/logout`.
...
  responses:
    Unauthenticated:
      description: Sin sesión válida (`UNAUTHENTICATED`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: La sesión es válida pero su rol no tiene permiso (`FORBIDDEN`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
    MedicalForbidden:
      description: |
        `FORBIDDEN` para ADMIN, RECEPCION o paciente; `PRACTITIONER_LINK_REQUIRED`
        para una cuenta MEDICO sin profesional asignado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
...
    UpsertSpecialtyRequest:
      type: object
      required: [name, description]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: Nombre recortado, con espacios colapsados y normalizado a minúsculas por el servidor.
        description:
          type: string
          minLength: 1
          maxLength: 500
    ScheduleBlockRequest:
      type: object
      required: [startAt, endAt]
      additionalProperties: false
      properties:
        startAt:
          type: string
          format: date-time
          description: ISO 8601 con hora de pared de `America/Guatemala` (UTC-06:00) cuando lo envía la UI.
        endAt:
          type: string
          format: date-time
          description: ISO 8601 con hora de pared de `America/Guatemala` (UTC-06:00) cuando lo envía la UI.
    ScheduleBlock:
      type: object
      required: [id, practitionerId, startAt, endAt, available]
      properties:
        id:
          type: string
        practitionerId:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        available:
          type: boolean
          description: False cuando el bloque ya tiene una cita activa; esos bloques siguen visibles al médico pero no al catálogo público.
    Specialty:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
          example: Odontología general
        description:
          type: string
...
    ApiError:
      type: object
      required: [status, code, message]
      properties:
        status:
          type: integer
          example: 400
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Datos inválidos
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                example: fecha
              message:
                type: string
                example: La fecha es obligatoria